Life_Expectancy_Analyzer
Life expectancy is the key metric for assessing population health. Based on  that I have collected data and tried to evaluate these figures in form of Line  chart in which we can set range (year) and compare growth among different  countries. For this I have used the following libraries of python: 
✓ Dash
 ✓ Plotly
 ✓ Pandas